Ignore Telnet options. Addresses issue #44.
If no option negotiation callback function is set, telnetlib
will pass Telnet options back to netconify in the data stream.
This is especially problematic for the AUTHENTICATION (0x25) option.
0x25 is an ASCII % character and confuses the login state machine into
thinking that it is at the shell prompt.
This function simply receives and ignores Telnet options. This prevents
the options from appearing in the data stream and confusing the
login state machine.